Create MissedRun.sh
This is to check if any job file missed its run. I have experienced that sometimes certain jobs go missing even after giving proper run commands. perhaps because of all the traffic at the clusters. This script will pick those job files, out of an entire ocean of job files in a directory, that somehow missed the run command, based on whether there are output files corresponding to those job files in that directory or not.